backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
openai==1.3.0
python-jose==3.3.0
sqlalchemy==2.0.23
pytest==7.4.3
black==23.11.0
flake8==6.1.0
diffusers==0.18.0
